Code review changes
1. Moved all the changes from contentactor to extendedcontentactor.
