+++ content-api/content-service/app/controllers/v4/ExtendedContentController.scala
@@ -92,4 +92,14 @@ class ExtendedContentController @Inject()(@Named(ActorNames.EXTENDED_CONTENT_ACT
     getResult(ApiId.CREATE_VERSION_CONTENT, contentActor, sbRequest)
   }
 
+  def extendedRead(identifier: String, mode: Option[String], fields: Option[String]) = Action.async { implicit request =>
+    val headers = commonReadHeaders()
+    val content = new java.util.HashMap().asInstanceOf[java.util.Map[String, Object]]
+    content.putAll(headers)
+    content.putAll(Map("identifier" -> identifier, "mode" -> mode.getOrElse("read"), "fields" -> fields.getOrElse("")).asJava)
+    val readRequest = getRequest(content, headers, "extendedReadContent")
+    setRequestContext(readRequest, version, objectType, schemaName)
+    getResult(ApiId.EXTENDED_READ_CONTENT, contentActor, readRequest, true)
+  }
+
 }
